+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### `jsonschema-generator`
+#### Added
+- enable look-up of annotations on a member's type parameter (e.g., a `Map`'s value type)
+- enable providing full custom schema definition to be included in `additionalProperties` or `patternProperties`
+
 #### Changed
 - bump `slf4j-api` dependency version from `1.7.35` to `2.0.3`
 - bump `jackson-core` dependency version from `2.13.2` to `2.13.4`